+++ CHANGELOG.md
@@ -16,7 +16,8 @@
 - Statistics can be reset by publishing specific MQTT messages to the `node-deepstackai-trigger/statistics` topics
   and sub-topics. Resolves [issue 308](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/306).
 - Statistics can be read and reset via new REST APIs. Overall statistics are available at `/statistics` and
-  per-trigger statistics are available at `/statistics/triggerName`. [See the API documentation for more information](https://github.com/danecreekphotography/node-deepstackai-trigger/wiki/REST-API). Resolves [issue 307](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/307).
+  per-trigger statistics are available at `/statistics/triggerName`. Statistics for all triggers can be retrieved from
+  `/statistics/all`. [See the API documentation for more information](https://github.com/danecreekphotography/node-deepstackai-trigger/wiki/REST-API). Resolves [issue 307](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/307) and [issue 311](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/311).
 - Shutting down the system after a failed launch no longer throws an error. Resolves [issue 301](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/291).
 - The underlying Linux variant used for the Docker image is now `node:slim`. Resolves [issue 299](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/291).