+++ CHANGELOG.md
@@ -18,6 +18,9 @@
 - Statistics can be read and reset via new REST APIs. Overall statistics are available at `/statistics` and
   per-trigger statistics are available at `/statistics/triggerName`. Statistics for all triggers can be retrieved from
   `/statistics/all`. [See the API documentation for more information](https://github.com/danecreekphotography/node-deepstackai-trigger/wiki/REST-API). Resolves [issue 307](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/307) and [issue 311](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/311).
+- Triggers can be activated by sending `node-deepstackai-trigger/motion` MQTT messages with the name of the trigger
+  in the message. This is similar to activating a trigger via the REST API and results in the trigger attempting
+  to download a snapshot from the `snapshotUri` specified in the trigger's configuration. Resolves [issue 314](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/291).
 - Shutting down the system after a failed launch no longer throws an error. Resolves [issue 301](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/291).
 - The underlying Linux variant used for the Docker image is now `node:slim`. Resolves [issue 299](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/291).
 - Pushbullet is now included in the settings JSON schema. Resolves [issue 316](https://github.com/danecreekphotography/node-deepstackai-trigger/issues/291).